Debug

+++ EveBot/EveBot.py
@@ -55,7 +55,7 @@ class Mycog:
                 results_split = results.split(' ')
                 sizem = results_split[11]
                 sizem = size.replace('m3','')
-                results_joined = ' '.join(results_split[1:4]) + ':\t' + "{:,}".format(float(results_split[9])) + ' isk\n' + ' '.join(results_split[4:7]) + ':\t' + "{:,}".format(float(results_split[10])) + ' isk\n' + ' '.join(results_split[7:9]) + ':\t' + "{:,}".format(float(sizem))+" m3"
+                results_joined = ' '.join(results_split[1:4]) + ':\t' + "{:,}".format(int(results_split[9])) + ' isk\n' + ' '.join(results_split[4:7]) + ':\t' + "{:,}".format(int(results_split[10])) + ' isk\n' + ' '.join(results_split[7:9]) + ':\t' + "{:,}".format(float(sizem))+" m3"
                 await self.bot.say('Results: \n' + results_joined)
             except:
                 await self.bot.say("Failed.")
